# Improve FITS header performance