--- a/astropy/io/fits/hdu/compressed.py
+++ b/astropy/io/fits/hdu/compressed.py
@@ -1471,6 +1471,10 @@ class CompImageHDU(BinTableHDU):
         if hasattr(self, '_image_header'):
             return self._image_header
 
+        if self._header is None and self._header_str is not None:
+            self._header = Header.fromstring(self._header_str)
+            self._header_str = None
+
         # Start with a copy of the table header.
         image_header = self._header.copy()