… CHANGES.rst
… astropy/io/fits/header.py