This PR is to optimize the initial parsing of FITS headers, as described in #5593 (which could probably be closed then). As discussed in #5593, a long-standing problem with `io.fits` is that to access a given HDU all the previous HDU headers must be parsed. And before the lazy loading feature, all HDU headers were parsed. So this is a crucial issue for FITS files with many HDUs, and often a barrier to use them.

The main issue is that the headers must be parsed to find the structural keywords (BITPIX, NAXIS*) that are needed to know the position of the following extension. After having tried to optimize the header parsing itself (#8428), which is always good, I came to the same conclusion as @embray (https://github.com/astropy/astropy/issues/5593#issuecomment-422493941): we need a fast header parser. Which this PR implements.  

The `_BasicHeader` class is added with a "fast" parser written in Cython. It is not as fast as I would like, only twice faster as the pure Python. I think that it is possible to do better but will keep this for the future, it is already much faster that `Header`. This new class is used only during the construction of the HDU classes, where we need only a subset of standard keywords, so COMMENT, HIERARCH, and special keywords are not parsed. The cards are parsed only if needed, and cached. Typically only a few cards will be parsed, during the HDU class detection step. Then the header string is stored, and used to construct the usual `Header` object if the HDU is accessed.

One remains that remains is to know if we need to parse to full header (currently implemented), or just the first block, which would bring another interesting speedup for headers with many cards (see https://github.com/astropy/astropy/issues/5593#issuecomment-290389892 and following comments). We could maybe imagine having a keyword to trigger the full header parsing ?

To give some numbers, and testing on different files:

`astropy/io/fits/tests/data/stddata.fits`, with 2 tables and just a few
keywords, to check if there is a penalty. The PR is still faster.

```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('astropy/io/fits/tests/data/stddata.fits', ext=1)"
```
*Before*: Mean +- std dev: 1.28 ms +- 0.03 ms
*After*: Mean +- std dev: 1.17 ms +- 0.02 ms

```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('astropy/io/fits/tests/data/stddata.fits', ext=2)"
```
*Before*: Mean +- std dev: 2.88 ms +- 0.10 ms
*After*: Mean +- std dev: 2.30 ms +- 0.08 ms

Then on a quite extreme case, a file with 335 extensions, where each header
has ~1700 cards (with HIERARCH):

```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=10)"
```
*Before*: Mean +- std dev: 173 ms +- 6 ms
*After*: Mean +- std dev: 26.8 ms +- 1.2 ms

```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=60)"
```
*Before*: Mean +- std dev: 1.01 sec +- 0.01 sec
*After*: Mean +- std dev: 101 ms +- 3 ms

```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=300)"
```
*Before*: Mean +- std dev: 5.26 sec +- 0.16 sec
*After*: Mean +- std dev: 450 ms +- 10 ms

![fits_perf_scipost](https://user-images.githubusercontent.com/311639/54467659-e4b88b00-4786-11e9-804b-78271f1eebe3.png)

And finally with a Euclid test file, with 650 extensions but a small number of keywords (15 to 20). It seems that they have more critical cases, and use `fitsio` instead of Astropy because of that.

![fits_perf_euclid](https://user-images.githubusercontent.com/311639/54467665-ee41f300-4786-11e9-8e41-41d040c43bf2.png)

ping @embray - if you have some time to have a look, it would be great to have your comments!