@MSeifert04 2019-03-18
Looks mostly good to me.

What bothers me is the performance, I would've expected it to be significantly faster, maybe it's all the string operations (slicing and joining) that slow it down. Did you do any profilings?

Since the tests are passing it's probably okay. One thing we really should do is add some docs on how this internally works.

Oh and I think some of the refactoring stuff would be good to do in separate PRs. Makes the changes a lot easier to review (for example changing the imports and the indentation of arguments/`.format` calls).