@saimn 2019-03-20
@MSeifert04 : 
> What bothers me is the performance, I would've expected it to be significantly faster, maybe it's all the string operations (slicing and joining) that slow it down. Did you do any profilings?

Yeah I would have expected a better performance as well, on the mockup I did it is much faster, and I don't know why. I tried several variants, keeping bytes, and nothing really helps. I think it is not critical for now, as even on small files it is faster, so I would prefer to move forward with this version and try to optimize later. Any idea or help to optimize will be greatly appreciated ;).

> Since the tests are passing it's probably okay. One thing we really should do is add some docs on how this internally works.

I don't think we currently have such docs? The additional complexity is rather limited compared to the mean complexity of io.fits :slightly_smiling_face:, the fast header is built for a short time, all in `_BaseHDU._readfrom_internal`, and its implementation is relatively straightforward.

> Oh and I think some of the refactoring stuff would be good to do in separate PRs. Makes the changes a lot easier to review (for example changing the imports and the indentation of arguments/`.format` calls).

I reverted the two import changes.