@mhvk 2019-03-23
@saimn - overall, this surely is the right way to go, but some questions. First, how much does the Cython buy you? It makes working on it harder for others, so it should be significant, yet since this is mostly string operations, it is not obvious to me that pre-compiling is going to help much.

Second, have you checked whether basing yourself on `abc.Mapping` is costly? I have no experience with this particular class, but did find that ABCs in general are not cheap: for every creation of an instance, it is checked whether the relevant methods are present (it cannot do this at the class creation level). Perhaps worth a try to just remove it?

Though really I guess some detailed profiling is needed - which perhaps is better done as follow up.