@MSeifert04 2019-03-24
> overall, this surely is the right way to go, but some questions. First, how much does the Cython buy you? It makes working on it harder for others, so it should be significant, yet since this is mostly string operations, it is not obvious to me that pre-compiling is going to help much.

Interesting point. I was assuming that most of the speedup is gained because of Cython. But you're right the potentially slow parts are string operations that will be (at least likely) equally fast in pure Python.

> Though really I guess some detailed profiling is needed - which perhaps is better done as follow up.

👍 - but I am not sure if we should do it as follow-up, since this is adding a lot of complexity and "only" gives a factor 2 improvement it should be essential that we're on the right track.