@mhvk 2019-03-24
It's true that most typical use cases with many extensions probably involve small headers, not the 1700 card ones that your PR really shines on... Also, your tests assume a user in fact knows which extension they want, where perhaps more realistically (some of) the other extensions also get looked at.  Is there currently a cache for extensions, so that, e.g., if I look for extension 80 after having looked at 70, one starts at 70 and not back at 1?