@astrofrog 2019-03-25
Just a quick note that it would probably make sense to add several benchmarks to https://github.com/astropy/astropy-benchmarks for this - maybe one simple case with one HDU, and then a couple of more complex examples?