@saimn 2019-03-25
@astrofrog - Yes, was working on it ;), see https://github.com/astropy/astropy-benchmarks/pull/80