@saimn 2019-03-25
@mhvk 
> overall, this surely is the right way to go, but some questions. First, how much does the Cython buy you? It makes working on it harder for others, so it should be significant, yet since this is mostly string operations, it is not obvious to me that pre-compiling is going to help much.

The pure python implementation is indeed already very fast. The cython implementation is faster only for files with many keywords, but this is not exceptional to have at least one header with a few thousands keywords, typically the PRIMARY. So the goal with the cython version is also to be sure that the overhead of _BasicHeader is negligible. Typically if you get the PRIMARY extension, it is first parsed by the _BasicHeader, and then Header will be build using the header string parsed by the _BasicHeader. So some work is duplicated (parsing the cards from the header string), but the file is not read twice, that's the reason of storing this header string.

It is difficult to give definitive numbers, as it depends on the headers, how many cards, if they use HIERARCH, or have many COMMENT, etc.
But to give an idea, on the Euclid file presented above, with very few cards, python and cython readers are equally fast. And on the other test files, with many keywords, the difference is significant.
This is before:
```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=60)"
Mean +- std dev: 132 ms +- 2 ms
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=300)"
Mean +- std dev: 613 ms +- 8 ms
```
and after : 
```
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=60)"
Mean +- std dev: 101 ms +- 3 ms
❯ python -m perf timeit --fast -s "from astropy.io import fits" "fits.getheader('scipost_white.fits', ext=300)"
Mean +- std dev: 450 ms +- 10 ms
```

Also from the benchmarks I just added (with many keywords), comparing the commit before cython and the head of the branch:
```
      21.4±0.03ms         21.1±0ms     0.98  io_fits.FITSHDUList.time_getheader
          198±3ms        157±0.1ms    ~0.79  io_fits.FITSHDUList.time_getheader_ext50
          316±1ms        256±0.6ms    ~0.81  io_fits.FITSHDUList.time_len
```

Also, it is only 50 lines of cython, the complexity is not huge. And the gain can be interesting. 

> Second, have you checked whether basing yourself on abc.Mapping is costly? I have no experience with this particular class, but did find that ABCs in general are not cheap: for every creation of an instance, it is checked whether the relevant methods are present (it cannot do this at the class creation level). Perhaps worth a try to just remove it?

I tested and cannot see a difference.